requirements.txt
python-dotenv==1.0.1
google-generativeai==0.8.4
sqlparse==0.5.3
fastapi==0.115.8
pydantic==2.10.6
slowapi==0.1.9
starlette==0.45.3
huggingface-hub==0.29.0
uvicorn==0.34.0
sentence-transformers==3.4.1
python-multipart==0.0.20
asyncpg==0.30.0